% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_filters.R
\name{apply_filters}
\alias{apply_filters}
\title{Apply filters}
\usage{
apply_filters(lattes.list, use.accepted)
}
\arguments{
\item{lattes.list}{a Lattes list object created using \code{\link[=lattes_to_list]{lattes_to_list()}}}

\item{use.accepted}{strategy to use with accepted papers}
}
\value{
Returns a filtered \emph{lattes.list}
}
\description{
Apply filters by Qualis / JCR / Year / UseAccepted
}
