% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_scores.R
\name{calculate_scores}
\alias{calculate_scores}
\title{Function to calculate the accreditation scores}
\usage{
calculate_scores(AllProductions, ignore.list = character(0))
}
\arguments{
\item{AllProductions}{data frame returned by \code{\link[=find_coauthorships]{find_coauthorships()}}}

\item{ignore.list}{list of full names to ignore when calculating scores.}
}
\value{
Updated \code{Allproductions} plus data frame with scores for each
professor.
}
\description{
Calculate the accreditation scores for PPGEE/UFMG
}
