% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_qualis_or_JCR.R
\name{filter_by_qualis_or_JCR}
\alias{filter_by_qualis_or_JCR}
\title{Filter a lattes.list by qualis extract / impact factor}
\usage{
filter_by_qualis_or_JCR(lattes.list, which.qualis = c("A1", "A2", "B1"),
  min.IF = 1.5)
}
\arguments{
\item{lattes.list}{Lattes list (generated internally in \code{\link[=lattes_to_list]{lattes_to_list()}})}

\item{which.qualis}{vector of desired Qualis levels}

\item{min.IF}{lower threshold for impact factor}
}
\value{
list containing parsed information on accepted journal papers
}
\description{
Keeps only papers from journals that are ranked into specific Qualis levels
or have impact factors greater than a given threshold.
}
