% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_year.R
\name{filter_by_year}
\alias{filter_by_year}
\title{Filter a lattes.list by year}
\usage{
filter_by_year(lattes.list, start.year = 0, end.year = 10000)
}
\arguments{
\item{lattes.list}{Lattes list (generated internally in \code{\link[=lattes_to_list]{lattes_to_list()}})}

\item{start.year}{initial year of the interval}

\item{end.year}{final year of the interval}
}
\value{
list containing parsed information on accepted journal papers
}
\description{
Removes papers that fall outside a given time window from a given
\code{lattes.list} object.
}
