% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_coauthorships.R
\name{find_coauthorships}
\alias{find_coauthorships}
\title{Function to find and tag duplicated entries (coauthorships)}
\usage{
find_coauthorships(filtered.list)
}
\arguments{
\item{filtered.list}{list of productions filtered by year, Qualis/JCR, and
publication status (accepted vs. published)}
}
\value{
Data frame summarizing the productions and co-authorship information
}
\description{
Find and tag duplicated entries (coauthorships) based on:
}
\details{
\enumerate{
\item DOI
\item Title + Journal Name + Journal Volume (using partial matching accepting
up to 5% differences in the aggregated names)
}
}
