% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_JCR.R
\name{set_JCR}
\alias{set_JCR}
\title{Obtain JCR information}
\usage{
set_JCR(lattes.list, JCR.file, min.IF, alias.list)
}
\arguments{
\item{lattes.list}{a Lattes list object created using \code{\link[=lattes_to_list]{lattes_to_list()}}}

\item{JCR.file}{CSV file containing the ranking of journals according to
the QUALIS classification system for a certain area. See \code{JCR File} for
details.}
}
\value{
Returns the original \emph{lattes.list} object updated with JCR information
for each paper
}
\description{
Attribute JCR information to each paper in \code{lattes.list}. The journal
information available in each entry of \code{lattes.list} is matched
against the information in \code{JCR.file}: first by ISSN and, if no match is
found or the ISSN is not informed in \code{lattes.list}, by name. Journal names
are reduced (removal of all spaces and punctuations) and upper-cased
before matching, to reduce the chances of name mismatches.
}
\section{JCR File}{

The JCR file must be a CSV file with commas as separators and UTF-8
encoding. To ensure these properties, follow the steps:
\itemize{
\item Obtain the relevant XLS file (e.g., from
\url{https://www.researchgate.net/post/New_Impact_factors_2017_for_Journals_are_released_now} or directly from the Clarivate Analytics Website
(if you have access). It is important to obtain the file containing ALL
listed journals.
\item Open the file using your reader of choice, remove the headers and footers
and copy the relevant information (including column names).
\item Open a new spreadsheet on Google Docs, \url{https://docs.google.com/spreadsheets/u/0/}
\item Paste everything into the Google Docs Spreadsheet
\item Download the resulting file as a comma-separated file (\strong{File} -> \strong{Download as} -> \strong{Comma-separated values (.csv, current sheet)})
}
}

