% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_qualis.R
\name{set_qualis}
\alias{set_qualis}
\title{Obtain QUALIS information}
\usage{
set_qualis(lattes.list, qualis.file, which.qualis)
}
\arguments{
\item{lattes.list}{a Lattes list object created using \code{\link[=lattes_to_list]{lattes_to_list()}}}

\item{qualis.file}{CSV file containing the ranking of journals according to
the QUALIS classification system for a certain area. See \code{Qualis File} for details.}

\item{which.qualis}{vector of valid qualis levels (to flag each paper's
\code{ValidQualis} entry)}
}
\value{
Returns the original \emph{lattes.list} object updated with Qualis information
for each paper
}
\description{
Attribute QUALIS information to each paper in \code{lattes.list}. The journal
information available in each entry of \code{lattes.list} is matched
against the information in \code{qualis.file}: first by ISSN and, if no match is
found or the ISSN is not informed in \code{lattes.list}, by name. Journal names
are reduced (removal of all spaces and punctuations) and upper-cased
before matching, to reduce the chances of name mismatches.
}
\section{Qualis File}{

The qualis file must be a CSV file with commas as separators and UTF-8
encoding. To ensure these properties, follow the steps:
\itemize{
\item Generate the relevant file from
\url{https://sucupira.capes.gov.br/sucupira/public/consultas/coleta/veiculoPublicacaoQualis/listaConsultaGeralPeriodicos.jsf},
selecting only the fields \emph{Evento de Classificacao} and \emph{Area de Avaliacao}
(leave the others blank).
\item Download the resulting \strong{xls} file.
\item Open the file using your reader of choice, select all and copy
\item Open a new spreadsheet on Google Docs, \url{https://docs.google.com/spreadsheets/u/0/}
\item Paste everything into the Google Docs Spreadsheet
\item Download the resulting file as a comma-separated file (\strong{File} -> \strong{Download as} -> \strong{Comma-separated values (.csv, current sheet)})
}
}

