% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_year.R
\name{set_year}
\alias{set_year}
\title{Set ValidYear flags}
\usage{
set_year(lattes.list, years)
}
\arguments{
\item{lattes.list}{a Lattes list object created using \code{\link[=lattes_to_list]{lattes_to_list()}}}

\item{years}{vector of start and end years,
\code{years = c(start.year, end.year)}.}
}
\value{
Returns the original \emph{lattes.list} object updated with the \code{ValidYear} flag
for each paper
}
\description{
Set the \code{ValidYear} flags for each paper in \code{lattes.list}, based on the
start and end years from \code{years}.
}
